# Reduce memory usage by replacing BTreeMap with BTreeSet 