Use chnlist.txt as test data and use GNU time to measure memory usage.

Testing environment: Win10 15063 WSL Ubuntu 16.04 x86_64 Linux 4.4.0-43-Microsoft
CPU: Intel Core i7-6700K

Empty IpRange: 1052k

Before (Using BTreeMap): 1508k

After (Using BTreeSet): 1252k

Benchmark:

Before:

```
running 7 tests
test create_ip_range_with_chnlists ... bench:   8,431,300 ns/iter (+/- 105,730)
test new_ip_range                  ... bench:       1,092 ns/iter (+/- 15)
test parse_chnlists                ... bench:     738,850 ns/iter (+/- 54,740)
test parse_one_ip                  ... bench:          21 ns/iter (+/- 0)
test parse_one_subnet1             ... bench:         105 ns/iter (+/- 0)
test parse_one_subnet2             ... bench:          95 ns/iter (+/- 1)
test test_10000_ips_in_chnlists    ... bench:   7,045,600 ns/iter (+/- 99,060)
```

After:

```
running 7 tests
test create_ip_range_with_chnlists ... bench:   8,237,000 ns/iter (+/- 609,450)
test new_ip_range                  ... bench:         592 ns/iter (+/- 8)
test parse_chnlists                ... bench:     742,300 ns/iter (+/- 5,820)
test parse_one_ip                  ... bench:          21 ns/iter (+/- 0)
test parse_one_subnet1             ... bench:         105 ns/iter (+/- 0)
test parse_one_subnet2             ... bench:          95 ns/iter (+/- 0)
test test_10000_ips_in_chnlists    ... bench:   6,896,900 ns/iter (+/- 132,480)
```